move Units.cs to own file
